misc1: Split out line_consumption
- Move all functions regarding physical line consumption into line_consumption.c
- Convert and add comments in doxygen style
- Create header line_consumption.h for inclusion by users of these functions
- Include this header in all files using plines_* functions
