+++ src/nvim/fold.c
@@ -26,6 +26,7 @@
 #include "nvim/memory.h"
 #include "nvim/message.h"
 #include "nvim/misc1.h"
+#include "nvim/line_consumption.h"
 #include "nvim/misc2.h"
 #include "nvim/garray.h"
 #include "nvim/move.h"
