+++ src/nvim/line_consumption.h
@@ -1,0 +1,7 @@
+#ifndef NVIM_LINE_CONSUMPTION_H
+#define NVIM_LINE_CONSUMPTION_H
+
+#ifdef INCLUDE_GENERATED_DECLARATIONS
+# include "line_consumption.h.generated.h"
+#endif
+#endif // NVIM_LINE_USAGE_H
