+++ src/nvim/misc1.c
@@ -1206,168 +1206,6 @@ int get_last_leader_offset(char_u *line, char_u **flags)
 }
 
 /*
- * Return the number of window lines occupied by buffer line "lnum".
- */
-int plines(linenr_T lnum)
-{
-  return plines_win(curwin, lnum, TRUE);
-}
-
-int 
-plines_win (
-    win_T *wp,
-    linenr_T lnum,
-    int winheight                  /* when TRUE limit to window height */
-)
-{
-  /* Check for filler lines above this buffer line.  When folded the result
-   * is one line anyway. */
-  return plines_win_nofill(wp, lnum, winheight) + diff_check_fill(wp, lnum);
-}
-
-int plines_nofill(linenr_T lnum)
-{
-  return plines_win_nofill(curwin, lnum, TRUE);
-}
-
-int 
-plines_win_nofill (
-    win_T *wp,
-    linenr_T lnum,
-    int winheight                  /* when TRUE limit to window height */
-)
-{
-  int lines;
-
-  if (!wp->w_p_wrap)
-    return 1;
-
-  if (wp->w_width == 0)
-    return 1;
-
-  /* A folded lines is handled just like an empty line. */
-  /* NOTE: Caller must handle lines that are MAYBE folded. */
-  if (lineFolded(wp, lnum) == TRUE)
-    return 1;
-
-  lines = plines_win_nofold(wp, lnum);
-  if (winheight > 0 && lines > wp->w_height)
-    return (int)wp->w_height;
-  return lines;
-}
-
-/*
- * Return number of window lines physical line "lnum" will occupy in window
- * "wp".  Does not care about folding, 'wrap' or 'diff'.
- */
-int plines_win_nofold(win_T *wp, linenr_T lnum)
-{
-  char_u      *s;
-  long col;
-  int width;
-
-  s = ml_get_buf(wp->w_buffer, lnum, FALSE);
-  if (*s == NUL)                /* empty line */
-    return 1;
-  col = win_linetabsize(wp, s, (colnr_T)MAXCOL);
-
-  /*
-   * If list mode is on, then the '$' at the end of the line may take up one
-   * extra column.
-   */
-  if (wp->w_p_list && lcs_eol != NUL)
-    col += 1;
-
-  /*
-   * Add column offset for 'number', 'relativenumber' and 'foldcolumn'.
-   */
-  width = wp->w_width - win_col_off(wp);
-  if (width <= 0)
-    return 32000;
-  if (col <= width)
-    return 1;
-  col -= width;
-  width += win_col_off2(wp);
-  return (col + (width - 1)) / width + 1;
-}
-
-/*
- * Like plines_win(), but only reports the number of physical screen lines
- * used from the start of the line to the given column number.
- */
-int plines_win_col(win_T *wp, linenr_T lnum, long column)
-{
-  long col;
-  char_u      *s;
-  int lines = 0;
-  int width;
-
-  /* Check for filler lines above this buffer line.  When folded the result
-   * is one line anyway. */
-  lines = diff_check_fill(wp, lnum);
-
-  if (!wp->w_p_wrap)
-    return lines + 1;
-
-  if (wp->w_width == 0)
-    return lines + 1;
-
-  s = ml_get_buf(wp->w_buffer, lnum, FALSE);
-
-  col = 0;
-  while (*s != NUL && --column >= 0) {
-    col += win_lbr_chartabsize(wp, s, (colnr_T)col, NULL);
-    mb_ptr_adv(s);
-  }
-
-  /*
-   * If *s is a TAB, and the TAB is not displayed as ^I, and we're not in
-   * INSERT mode, then col must be adjusted so that it represents the last
-   * screen position of the TAB.  This only fixes an error when the TAB wraps
-   * from one screen line to the next (when 'columns' is not a multiple of
-   * 'ts') -- webb.
-   */
-  if (*s == TAB && (State & NORMAL) && (!wp->w_p_list || lcs_tab1))
-    col += win_lbr_chartabsize(wp, s, (colnr_T)col, NULL) - 1;
-
-  /*
-   * Add column offset for 'number', 'relativenumber', 'foldcolumn', etc.
-   */
-  width = wp->w_width - win_col_off(wp);
-  if (width <= 0)
-    return 9999;
-
-  lines += 1;
-  if (col > width)
-    lines += (col - width) / (width + win_col_off2(wp)) + 1;
-  return lines;
-}
-
-int plines_m_win(win_T *wp, linenr_T first, linenr_T last)
-{
-  int count = 0;
-
-  while (first <= last) {
-    int x;
-
-    /* Check if there are any really folded lines, but also included lines
-     * that are maybe folded. */
-    x = foldedCount(wp, first, NULL);
-    if (x > 0) {
-      ++count;              /* count 1 for "+-- folded" line */
-      first += x;
-    } else {
-      if (first == wp->w_topline)
-        count += plines_win_nofill(wp, first, TRUE) + wp->w_topfill;
-      else
-        count += plines_win(wp, first, TRUE);
-      ++first;
-    }
-  }
-  return count;
-}
-
-/*
  * Insert string "p" at the cursor position.  Stops at a NUL byte.
  * Handles Replace mode and multi-byte characters.
  */
