+++ src/nvim/move.c
@@ -26,6 +26,7 @@
 #include "nvim/mbyte.h"
 #include "nvim/memline.h"
 #include "nvim/misc1.h"
+#include "nvim/line_consumption.h"
 #include "nvim/misc2.h"
 #include "nvim/popupmnu.h"
 #include "nvim/screen.h"
