+++ src/nvim/normal.c
@@ -39,6 +39,7 @@
 #include "nvim/memory.h"
 #include "nvim/message.h"
 #include "nvim/misc1.h"
+#include "nvim/line_consumption.h"
 #include "nvim/misc2.h"
 #include "nvim/keymap.h"
 #include "nvim/move.h"
