+++ src/nvim/screen.c
@@ -110,6 +110,7 @@
 #include "nvim/menu.h"
 #include "nvim/message.h"
 #include "nvim/misc1.h"
+#include "nvim/line_consumption.h"
 #include "nvim/misc2.h"
 #include "nvim/garray.h"
 #include "nvim/move.h"
